runners/azure-blob-eventgrid-detect/requirements.txt
azure-data-tables>=12.6,<13
azure-identity>=1.17,<2
azure-servicebus>=7.13,<8
azure-storage-blob>=12.24,<13
